Move IM §9 v21-specific classes from .common
